Rename some behaviors with more obvious names.
