+++ crossdock/client/errors_httpclient/behavior.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package errors
+package errors_httpclient
 
 import (
 	"fmt"
