+++ crossdock/client/tchclient/behavior.go
@@ -36,7 +36,7 @@ const (
 
 var log = tchannel.SimpleLogger
 
-// Run executes the tchclient test
+// Run exercises a YARPC server from a tchannel client.
 func Run(t crossdock.T) {
 	fatals := crossdock.Fatals(t)
