+++ docker-compose.yml
@@ -17,10 +17,10 @@ services:
             - AXIS_SERVER=go,node,java,python
             - AXIS_TRANSPORT=http,tchannel
             - AXIS_ENCODING=raw,json,thrift
-            - AXIS_ERRORSHTTPOUT=node,go
-            # AXIS_ERRORSHTTPIN TODO
-            - AXIS_ERRORSTCHOUT=go
-            # AXIS_ERRORSTCHIN TODO
+            - AXIS_ERRORS_HTTPCLIENT=node,go
+            # AXIS_ERRORS_HTTPSERVER TODO
+            - AXIS_ERRORS_TCHCLIENT=go
+            # AXIS_ERRORS_TCHSERVER TODO
             - AXIS_CTXCLIENT=go
             - AXIS_CTXSERVER=go
             - AXIS_GAUNTLET=go
@@ -30,9 +30,9 @@ services:
             - BEHAVIOR_JSON=client,server,transport
             - BEHAVIOR_THRIFT=client,server,transport
             - BEHAVIOR_HEADERS=client,server,transport,encoding
-            - BEHAVIOR_ERRORS=errorshttpout,server
+            - BEHAVIOR_ERRORS_HTTPCLIENT=errors_httpclient,server
             # BEHAVIOR_ERRORSHTTPIN TODO
-            - BEHAVIOR_ERRORSTCHOUT=errorstchout,server
+            - BEHAVIOR_ERRORS_TCHCLIENT=errors_tchclient,server
             # BEHAVIOR_ERRORSTCHIN TODO
             - BEHAVIOR_TCHCLIENT=client,server,encoding
             - BEHAVIOR_TCHSERVER=client,server,encoding
